# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "ACROSS Server",
        "summary": "Astrophysics Cross-Observatory Science Support (ACROSS)",
        "description": "Server providing tools and utilities for various NASA missions to aid in coordination of large observation efforts.",
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/permission/": {
            "get": {
                "tags": [
                    "Permission"
                ],
                "summary": "Read permissions",
                "description": "Read many permissions.",
                "operationId": "get_permissions",
                "responses": {
                    "200": {
                        "description": "A list of permissions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/Permission"
                                    },
                                    "type": "array",
                                    "title": "Response 200 Get Permissions"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The permission does not exist."
                    }
                }
            }
        },
...
        "/user/{user_id}/group/{group_id}": {
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Leave a group",
                "description": "Leave a group by id",
                "operationId": "leave_group",
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "UUID of the user"
                        }
                    },
                    {
                        "name": "group_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Group Id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successfully left the group"
                    },
                    "404": {
                        "description": "The user does not exist."
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/group/": {
            "get": {
                "tags": [
                    "Group"
                ],
                "summary": "Read groups",
                "description": "Read many groups.",
                "operationId": "get_groups",
                "responses": {
                    "200": {
                        "description": "A list of groups",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/across_server__routes__group__schemas__Group"
                                    },
                                    "type": "array",
                                    "title": "Response 200 Get Groups"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The group does not exist."
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/group/{group_id}": {
            "get": {
                "tags": [
                    "Group"
                ],
                "summary": "Read a group",
                "description": "Read a group by role ID.",
                "operationId": "get_group",
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "parameters": [
                    {
                        "name": "group_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "UUID of the group"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The requested group",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/across_server__routes__group__schemas__Group"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The group does not exist."
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/group/{group_id}/user/{user_id}": {
            "delete": {
                "tags": [
                    "Group"
                ],
                "summary": "Remove a user from a group",
                "description": "Remove a user from a group by ID",
                "operationId": "remove_user",
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "User Id"
                        }
                    },
                    {
                        "name": "group_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "UUID of the group"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "404": {
                        "description": "The group does not exist."
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "GroupRoleRead": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "permissions": {
                        "items": {
                            "$ref": "#/components/schemas/Permission"
                        },
                        "type": "array",
                        "title": "Permissions"
                    },
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "permissions",
                    "id"
                ],
                "title": "GroupRoleRead"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "Permission": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "name"
                ],
                "title": "Permission"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
            "across_server__routes__group__role__schemas__GroupRole": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "permissions": {
                        "items": {
                            "$ref": "#/components/schemas/Permission"
                        },
                        "type": "array",
                        "title": "Permissions"
                    },
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "users": {
                        "items": {
                            "$ref": "#/components/schemas/across_server__routes__group__role__schemas__User"
                        },
                        "type": "array",
                        "title": "Users"
                    },
                    "service_accounts": {
                        "anyOf": [
                            {
                                "items": {
                                    "$ref": "#/components/schemas/across_server__routes__group__role__schemas__ServiceAccount"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Service Accounts"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "permissions",
                    "id",
                    "users",
                    "service_accounts"
                ],
                "title": "GroupRole"
            },
            "across_server__routes__group__role__schemas__ServiceAccount": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "description": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Description"
                    },
                    "user": {
                        "$ref": "#/components/schemas/across_server__routes__group__role__schemas__User"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "name",
                    "description",
                    "user"
                ],
                "title": "ServiceAccount"
            },
            "across_server__routes__group__role__schemas__User": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "first_name": {
                        "type": "string",
                        "title": "First Name"
                    },
                    "last_name": {
                        "type": "string",
                        "title": "Last Name"
                    },
                    "username": {
                        "type": "string",
                        "title": "Username"
                    },
                    "email": {
                        "type": "string",
                        "title": "Email"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "first_name",
                    "last_name",
                    "username",
                    "email"
                ],
                "title": "User"
            },
            "across_server__routes__group__schemas__Group": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "short_name": {
                        "type": "string",
                        "title": "Short Name"
                    },
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "users": {
                        "items": {
                            "$ref": "#/components/schemas/across_server__routes__group__schemas__User"
                        },
                        "type": "array",
                        "title": "Users"
                    },
                    "roles": {
                        "items": {
                            "$ref": "#/components/schemas/across_server__routes__group__role__schemas__GroupRole"
                        },
                        "type": "array",
                        "title": "Roles"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "short_name",
                    "id",
                    "users",
                    "roles"
                ],
                "title": "Group"
            },
            "across_server__routes__group__schemas__User": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "username": {
                        "type": "string",
                        "title": "Username"
                    },
                    "first_name": {
                        "type": "string",
                        "title": "First Name"
                    },
                    "last_name": {
                        "type": "string",
                        "title": "Last Name"
                    },
                    "email": {
                        "type": "string",
                        "title": "Email"
                    },
                    "group_roles": {
                        "items": {
                            "$ref": "#/components/schemas/GroupRoleRead"
                        },
                        "type": "array",
                        "title": "Group Roles"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "username",
                    "first_name",
                    "last_name",
                    "email",
                    "group_roles"
                ],
                "title": "User"
            },
...
        },
        "securitySchemes": {
            "Authorization": {
                "type": "http",
                "description": "Enter your access token.",
                "scheme": "bearer"
            }
        }
    }
}